+++ pkg/checks/audit/audit.go
@@ -79,6 +79,7 @@ func (d Deps) client(ctx context.Context) (kubernetes.Interface, error) {
 func init() {
 	checks.Register(ExemptionsCommand())
 	checks.Register(HardeningCommand(Deps{}))
+	checks.Register(NetpolCommand(Deps{}))
 	checks.Register(WorkloadsCommand(Deps{}))
 }